apps/server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fluxora-server"
version = "0.1.0"
description = "Fluxora server backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "aiosqlite==0.20.0",
    "zeroconf==0.131.0",
    "aiortc==1.9.0",
    "httpx==0.27.0",
    "python-multipart==0.0.9",
    "pydantic-settings==2.14.0",
    "slowapi==0.1.9",
    "python-json-logger==4.1.0",
    "keyring==25.7.0",
    "cryptography==47.0.0",
    "argon2-cffi==25.1.0",
    "psutil==7.2.2"
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.7",
    "httpx==0.27.0",
    "black==24.4.0",
    "ruff==0.15.12"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["routers*", "models*", "services*", "utils*", "database*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
